+++ clients/privacy-center/types/api/models/DBActions.ts
@@ -1,0 +1,12 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * The available path parameters for the `/admin/db/{action}` endpoint.
+ */
+export enum DBActions {
+  UPGRADE = "upgrade",
+  RESET = "reset",
+  DOWNGRADE = "downgrade",
+}