+++ clients/privacy-center/types/api/models/IdentityTypes.ts
@@ -1,0 +1,8 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type IdentityTypes = {
+  email: boolean;
+  phone_number: boolean;
+};