+++ clients/privacy-center/types/api/models/Page_ConsentReportingSchema_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { ConsentReportingSchema } from "./ConsentReportingSchema";
+
+export type Page_ConsentReportingSchema_ = {
+  items: Array<ConsentReportingSchema>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};