+++ clients/privacy-center/types/api/models/Page_MessagingTemplateWithPropertiesSummary_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { MessagingTemplateWithPropertiesSummary } from "./MessagingTemplateWithPropertiesSummary";
+
+export type Page_MessagingTemplateWithPropertiesSummary_ = {
+  items: Array<MessagingTemplateWithPropertiesSummary>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};