+++ clients/privacy-center/types/api/models/Page_PreApprovalWebhookResponse_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { PreApprovalWebhookResponse } from "./PreApprovalWebhookResponse";
+
+export type Page_PreApprovalWebhookResponse_ = {
+  items: Array<PreApprovalWebhookResponse>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};