+++ clients/privacy-center/types/api/models/Page_StagedResourceAPIResponse_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { StagedResourceAPIResponse } from "./StagedResourceAPIResponse";
+
+export type Page_StagedResourceAPIResponse_ = {
+  items: Array<StagedResourceAPIResponse>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};