+++ clients/privacy-center/types/api/models/ResourceTypes.ts
@@ -1,0 +1,14 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * Resource types that can use custom fields.
+ */
+export enum ResourceTypes {
+  SYSTEM = "system",
+  DATA_USE = "data use",
+  DATA_CATEGORY = "data category",
+  DATA_SUBJECT = "data subject",
+  PRIVACY_DECLARATION = "privacy declaration",
+}