+++ clients/privacy-center/types/api/models/TCFSpecialPurposeRecord.ts
@@ -29,7 +29,7 @@ export type TCFSpecialPurposeRecord = {
    * The fideslang default taxonomy data uses that are associated with the purpose.
    */
   data_uses: Array<string>;
-  default_preference?: UserConsentPreference;
+  default_preference?: UserConsentPreference | null;
   vendors?: Array<EmbeddedVendor>;
   systems?: Array<EmbeddedVendor>;
   legal_bases?: Array<string>;